chore(ci): :green_heart: Refine GitHub Actions workflow for mkdocs-material installation
This commit enhances the GitHub Actions workflow by updating the conditions for installing the `mkdocs-material` package based on the presence of the `GH_TOKEN` secret and the event type. The installation commands are modified to remove the `--system` flag, ensuring a cleaner environment setup. Additionally, the conditions for deploying documentation to the `dev` and `master` branches are refined to improve clarity and functionality. These changes aim to streamline the CI/CD process and ensure more reliable documentation builds.