ci(makefile): :wrench: Enhance testing commands and add notebook support
This commit updates the Makefile to improve testing commands by adding new options for running Jupyter notebook tests and multi-version tests with tox. The `test-all` command is modified to include notebook tests, and new commands `test-notebooks`, `test-notebook`, and `test-all-notebooks` are introduced for better notebook testing management. Additionally, the installation of `nbmake` is included in the test tools setup, enhancing the project's testing capabilities.